requirement.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.6
ndg-httpsclient==0.5.1
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.5.1
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
